requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
openai>=1.13.0
python-dotenv>=1.0.0
reportlab>=4.0.0

# File processing libraries for document uploads
PyPDF2>=3.0.0
python-docx>=1.1.0
docx2txt>=0.9
pdfplumber>=0.11.0
radon>=6.0.1
flake8>=7.1.1